Fix various feedback from PR 1930
