+++ include/ttmlir/Dialect/TTIR/IR/TTIRTileOps.h
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#ifndef TTMLIR_DIALECT_TTIR_IR_TTIRTILEOPS_H
+#define TTMLIR_DIALECT_TTIR_IR_TTIRTILEOPS_H
+
+#include "ttmlir/Dialect/TTIR/IR/TTIROps.h"
+
+#define GET_OP_CLASSES
+#include "ttmlir/Dialect/TTIR/IR/TTIRTileOps.h.inc"
+
+#endif // TTMLIR_DIALECT_TTIR_IR_TTIRTILEOPS_H
