+++ lib/Dialect/TTIR/IR/CMakeLists.txt
@@ -3,6 +3,7 @@ add_mlir_dialect_library(MLIRTTIRDialect
         TTIROps.cpp
         TTIROpsInterfaces.cpp
         TTIRTraits.cpp
+        TTIRTileOps.cpp
 
         ADDITIONAL_HEADER_DIRS
         ${PROJECT_SOURCE_DIR}/include/ttmlir
