+++ lib/Dialect/TTIR/IR/TTIROpsInterfaces.cpp
@@ -49,18 +49,3 @@ mlir::tt::ttir::detail::verifyBroadcastable(mlir::Operation *op) {
 
   return success();
 }
-
-mlir::LogicalResult
-mlir::tt::ttir::detail::verifyGenericParent(mlir::Operation *op) {
-  mlir::Operation *parent = op->getParentOp();
-
-  while (parent) {
-    if (llvm::dyn_cast<ttir::GenericOp>(parent) ||
-        llvm::dyn_cast<linalg::GenericOp>(parent)) {
-      return success();
-    }
-    parent = op->getParentOp();
-  }
-
-  return op->emitOpError("TTIR Tile Ops must be inside a generic region");
-}
